requirements.txt
pandas
tabulate
pywin32
python-dotenv
pymongo